moved data wrapping to the container
